// src/api/invoice/tags.ts
export enum Tags {
    DEL_INVOICE = 'DEL_INVOICE',
    GET_INVOICES_ISSUED = 'GET_INVOICES_ISSUED',
    GET_INVOICE = 'GET_INVOICE',
    GET_COMPANY_INVOICES = 'GET_COMPANY_INVOICES',
    POST_INVOICE = 'POST_INVOICE',
    POST_CALCULATE_INVOICE_PRICE = 'POST_CALCULATE_INVOICE_PRICE',
    POST_SEND_INVOICE_EMAIL = 'POST_SEND_INVOICE_EMAIL',
    PUT_INVOICE = 'PUT_INVOICE',
    PATCH_MARK_AS_PAID_INVOICE = 'PATCH_MARK_AS_PAID_INVOICE',
    POST_REQUEST_EARLY_PAYMENT = 'POST_REQUEST_EARLY_PAYMENT',
    POST_VALIDATE_USER_COMPANY = 'POST_VALIDATE_USER_COMPANY'
}